+++ govtool/frontend/src/components/organisms/HomeCards.tsx
@@ -34,7 +34,7 @@ export const HomeCards = () => {
 
   const onClickLearnMoreAboutDRepRegistration = () =>
     openInNewTab(
-      "https://docs.sanchogov.tools/faqs/what-does-it-mean-to-register-as-a-drep",
+      "https://docs.sanchogov.tools/how-to-use-the-govtool/using-govtool/dreps",
     );
 
   const onClickLearnMoreAboutDirectVoterRegistration = () =>
@@ -44,7 +44,7 @@ export const HomeCards = () => {
 
   const onClickLearnMoreAboutProposingGovAction = () =>
     openInNewTab(
-      "https://docs.sanchogov.tools/how-to-use-the-govtool/using-govtool/propose-a-governance-action",
+      "https://docs.sanchogov.tools/how-to-use-the-govtool/using-govtool/governance-actions/propose-a-governance-action",
     );
 
   const navigateToGovActions = useCallback(