+++ govtool/frontend/src/services/requests/getDRepList.ts
@@ -1,5 +1,11 @@
-import type { InfinityDRepData, DRepStatus, DRepListSort } from "@models";
+import {
+  type InfinityDRepData,
+  type DRepStatus,
+  type DRepListSort,
+  MetadataStandard,
+} from "@models";
 import { API } from "../API";
+import { postValidate } from "./metadataValidation";
 
 export type GetDRepListArguments = {
   filters?: string[];
@@ -18,7 +24,7 @@ export const getDRepList = async ({
   searchPhrase = "",
   status = [],
 }: GetDRepListArguments): Promise<InfinityDRepData> => {
-  const response = await API.get("/drep/list", {
+  const response = await API.get<InfinityDRepData>("/drep/list", {
     params: {
       page,
       pageSize,
@@ -28,5 +34,34 @@ export const getDRepList = async ({
       ...(status.length && { status }),
     },
   });
-  return response.data;
+
+  const validatedResponse = {
+    ...response.data,
+    elements: await Promise.all(
+      response.data.elements.map(async (drep) => {
+        if (drep.metadataStatus || drep.metadataValid) {
+          return drep;
+        }
+        if (drep.url && drep.metadataHash) {
+          const validationResponse = await postValidate({
+            url: drep.url,
+            hash: drep.metadataHash,
+            standard: MetadataStandard.CIPQQQ,
+          });
+          return {
+            ...drep,
+            bio: validationResponse.metadata?.bio,
+            dRepName: validationResponse.metadata?.dRepName,
+            email: validationResponse.metadata?.email,
+            references: validationResponse.metadata?.references,
+            metadataStatus: validationResponse.status || null,
+            metadataValid: validationResponse.valid,
+          };
+        }
+        return drep;
+      }),
+    ),
+  };
+
+  return validatedResponse;
 };