+++ govtool/frontend/src/services/requests/getDRepVotes.ts
@@ -1,5 +1,6 @@
-import { VotedProposal } from "@models";
+import { MetadataStandard, VotedProposal } from "@models";
 import { API } from "../API";
+import { postValidate } from "./metadataValidation";
 
 type GetDRepVotesParams = {
   type?: string[];
@@ -18,5 +19,38 @@ export const getDRepVotes = async ({
 
   const { data } = await API.get<VotedProposal[]>(urlBase, { params });
 
-  return data;
+  const validatedData = await Promise.all(
+    data.map(async (votedProposal) => {
+      if (
+        votedProposal.proposal.metadataStatus ||
+        votedProposal.proposal.metadataValid
+      ) {
+        return votedProposal;
+      }
+
+      if (votedProposal.proposal.url && votedProposal.proposal.metadataHash) {
+        const validationResponse = await postValidate({
+          url: votedProposal.proposal.url,
+          hash: votedProposal.proposal.metadataHash,
+          standard: MetadataStandard.CIP108,
+        });
+
+        return {
+          ...votedProposal,
+          proposal: {
+            ...votedProposal.proposal,
+            abstract: validationResponse.metadata?.abstract,
+            motivation: validationResponse.metadata?.motivation,
+            title: validationResponse.metadata?.title,
+            rationale: validationResponse.metadata?.rationale,
+            references: validationResponse.metadata?.references,
+            metadataStatus: validationResponse.status || null,
+            metadataValid: validationResponse.valid,
+          },
+        };
+      }
+    }),
+  );
+
+  return validatedData;
 };